Refaktoryzacja App.js i utworzenie kontekstu aplikacji
